+++ CHANGES.rst
@@ -2,6 +2,8 @@
 ====================
 
 - added :meth:`udev.Enumerator.match_toplevel` to find top-level devices
+- ``>``, ``>=``, ``<`` or ``<=`` raise :exc:`~exceptions.TypeError` now, if
+  used on :class:`udev.Device` objects.
 
 
 0.2 (Jun 28, 2010)